DCMAW-5593: Move utils into k6
