+++ deploy/scripts/src/mobile/README.md
