+++ deploy/scripts/src/mobile/frontend-e2e.test.ts
@@ -30,8 +30,8 @@ const profiles: ProfileList = {
       preAllocatedVUs: 75,
       maxVUs: 120,
       stages: [
-        { target: 5, duration: '30s' },
-        { target: 5, duration: '30s' }
+        { target: 5, duration: '10s' },
+        // { target: 5, duration: '30s' }
       ],
       exec: 'mamIphonePassport'
     }
